General Attack Base Properties Updates
Change Hit rate from float -> Int
Change to update Stun/Stop values for command normals & string attack normals
Changed Attack Hashes to singular int rather than list of ints

+++ Project Strongest/Assets/FGCMaker/Scripts/CharacterBasedCode/NewCharacterController/AttackHandler_Attack.cs
@@ -428,7 +428,7 @@ public class AttackHandler_Attack : AttackHandler_Base
 public class HitCount
 {
     public int _count, _startCount;
-    [Range(0, 30)] public float _refreshRate, _startRefreshRate;
+    [Range(0, 30)] public int _refreshRate, _startRefreshRate;
     public void ResetRefresh()
     {
         _refreshRate = _startRefreshRate;
